Remove unnecessary dark mode classes from components
The dark mode classes have been removed from various components including Hero, Footer, HomePreview, and Header components among others. This refactoring was necessary as it simplifies the code and removes the elements that were not contributing to the overall functionality. The removed classes were initially used for setting styles in dark mode, but the mode is not currently being used, hence the cleanup. This change makes the code more maintain