Refactor page query and CardBlock component in datacms.query.ts and CardBlock.tsx
This revision simplifies the PAGE_QUERY in datacms.query.ts by removing unused fields, and adapts for changes in CardBlock component properties. The price field is removed from CardBlockRecord and CardBlockProps since it was not effectively used and was causing confusion with erroneous data. New fields are added to provide more information about the cards. This will enhance the page load performance and enhance maintainability of code.