Optimize data query and modify CardBlock component
The code refactor was implemented to remove the 'price' field in PAGE_QUERY in datacms.query.ts and in the CardBlock component properties. This field was not actually being used and its presence may lead to data inconsistency issues. Moreover, extra optional chaining has been added in the map function in the rendering logic of page.tsx for better null safety. This ensures that the page does not crash when 'home' field is not defined in data object. These changes enhance the page's load speed and significantly improves the readability and maintainability of the code.