Make Card component's quote and job props optional
The quote and job props of the Card component were changed to optional to increase its flexibility and accommodate instances where these data might not be available. Now, the 'quote' section will only render if a quote exists and the 'job' line was commented out for possible future use. These changes improve the robustness and versatility of the Card component.