Remove 'job' prop from Card component
The 'job' prop of the Card component was removed to increase the component's versatility as it is not always applicable or available. Given this, the 'job' attribute was also removed from the datacms.query to ensure data fetched is consistent with the updated Card component. The line displaying 'job' on the Card was commented out instead of deleted, in case there is a future need to reintroduce it.